Refactor RZ select and clean legacy files
