+++ src/services/importer.js
@@ -1,34 +1,5 @@
-// src/services/importer.js
 import { parsePlanilha } from '../utils/excel.js';
 import { addLot, setCurrentLotId, addItemsBulk } from '../store/db.js';
-
-const META_KEY = 'confApp.meta.v1';
-
-export function saveMeta(partial) {
-  const cur = loadMeta();
-  const next = { ...cur, ...partial, savedAt: new Date().toISOString() };
-  localStorage.setItem(META_KEY, JSON.stringify(next));
-  return next;
-}
-
-export function loadMeta() {
-  try { return JSON.parse(localStorage.getItem(META_KEY) || '{}'); }
-  catch { return {}; }
-}
-
-export function wireLotFileCapture(inputEl) {
-  inputEl?.addEventListener('change', () => {
-    const file = inputEl.files?.[0];
-    if (file) saveMeta({ loteName: file.name });
-  });
-}
-
-export function wireRzCapture(selectEl) {
-  selectEl?.addEventListener('change', () => {
-    const rz = selectEl.value || '';
-    saveMeta({ rz });
-  });
-}
 
 function mapRowToItem(row, lotId) {
   return {
